Add map type button to map in route state
